requirements.txt
numpy~=1.19.2
scipy~=1.6.0
pandas
tqdm
cbox~=0.5.0
scikit-learn~=0.23.2
Pillow~=8.1.0
imageio
pygifsicle

torchvision~=0.8.2
torch

pytest~=6.2.2
